requirements.txt
pandas
numpy
scipy
Cython
Pillow
scikit-learn
tensorflow-gpu
keras
matplotlib
tqdm
seaborn
